requirements.txt
amqp==5.1.1
anyio==4.0.0
async-timeout==4.0.3
attrs==23.1.0
billiard==4.1.0
blinker==1.6.3
Brotli==1.1.0
celery==5.3.4
certifi==2023.7.22
charset-normalizer==3.3.0
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
ConfigArgParse==1.7
exceptiongroup==1.1.3
Faker==19.6.2
fastapi==0.95.0
flake8==6.1.0
Flask==3.0.0
Flask-BasicAuth==0.2.0
Flask-Cors==4.0.0
flower==1.2.0
gevent==23.9.1
geventhttpclient==2.0.11
greenlet==3.0.0
h11==0.14.0
httpcore==0.18.0
httpx==0.25.0
humanize==4.8.0
idna==3.4
importlib-metadata==6.8.0
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
kombu==5.3.2
locust==2.16.1
MarkupSafe==2.1.3
mccabe==0.7.0
msgpack==1.0.7
mysql-connector-python==8.1.0
nodeenv==1.8.0
packaging==23.2
pluggy==1.3.0
prisma==0.10.0
prometheus-client==0.17.1
prompt-toolkit==3.0.39
protobuf==4.21.12
psutil==5.9.5
pycodestyle==2.11.0
pydantic==1.10.13
pyflakes==3.1.0
pytest==7.2.2
pytest-celery==0.0.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python-multipart==0.0.6
pytz==2023.3.post1
pyzmq==25.1.1
redis==4.5.4
requests==2.31.0
roundrobin==0.0.4
six==1.16.0
sniffio==1.3.0
starlette==0.26.1
tomli==2.0.1
tomlkit==0.12.1
tornado==6.3.3
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.0.6
uvicorn==0.21.1
vine==5.0.0
wcwidth==0.2.8
Werkzeug==3.0.0
zipp==3.17.0
zope.event==5.0
zope.interface==6.1
